Find previous value of a field from Notes history PPM
PPM logs values of a field if notes history is enabled. When you want to report the previous value of a field use this query. You need to outer join if you want to display current values. That is, Create your report and than outer join with the query below to show current and the last value.